% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloFromSP.R
\name{phyloFromSP}
\alias{phyloFromSP}
\title{Create phylogeny from a 'species-parent' matrix with origin and extinction times}
\usage{
phyloFromSP(sp)
}
\arguments{
\item{sp}{a 'species-parent' matrix as described in the details}
}
\value{
A phylogeny with tip labels and edge length are derived from the input matrix
}
\description{
Create a phylogeny from a matrix giving species and ancestry information.
}
\details{
An 'sp' matrix is a matrix where the first column is the species ID, the 
second column is the parent species, the third is the origin time of the species indicated by
the first column, and the forth column is that species extinction time
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
