% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDiv.R
\name{simDiv}
\alias{simDiv}
\title{Simulate diversification on an adaptive landscape}
\usage{
simDiv(la, mu, r, d, x0, opt0, bsd, ngen)
}
\arguments{
\item{la}{function of position in the landscape giving birth rate at that point}

\item{mu}{function of position in the landscape giving death rate at that point}

\item{r}{the rate of evolutionary movement in the landscape}

\item{d}{the critical distance past which a new species is generated}

\item{x0}{initial trait value of the initializing population}

\item{opt0}{initial optimal value in the adaptive landscape}

\item{bsd}{SD of the Brownian drift of the adaptive optimum}

\item{ngen}{number of generations over which to run the simulation}
}
\value{
A phylogeny with tip labels and edge length are derived from the input matrix
}
\description{
Simulate diversification on an adaptive landscape
}
\details{
The adaptiveness of the landscape is determined by the functions \code{la} and \code{mu}
which capture how speciation and extinction change as a function of position in the landscape
la, mu, r, d, x0, opt0, bsd, ngen
}
\author{
Andy Rominger <ajrominger@gmail.com>
}
